undo template.go for elasticsearch
